AVI/LE/HTTP-01: Add a bootstrapping feature.
As it stands, Voyager cannot handle bootstrapping full-SSL ingresses on its own.
This automation performs all the exact steps described in Voyager's documentation
(https://appscode.com/products/voyager/10.0.0/guides/certificate/http/overview/)
as in: bringing up a non-TLS ingress, using it for the certificate, enabling TLS
when the certificate is ready. We're using Kubernetes facts to understand when our
Secret is ready to be consumed.

This is done only once upon first deployment of the certificate: it is not needed
for renewals and it is literally required only when bootstrapping the Ingress.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>