AVI/API: Move to HAProxy backend rules for rewriting targets
When using `ingress.appscode.com/rewrite-target`, the rewrite rule is applied to *every* single routing
declared in the ingress - this also includes the ACME challenge. In fact, after the rewrite the challenge
can never be solved due to the fact that the Solver receives bogus parameters.

Rather than using the rule at the ingress-level, use the not-so-pretty native HAProxy backend rule,
which can be individually applied to each backend. This way, the http-01 challenge goes through.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>