AVI/Certificate: Use a safer bootstrap mechanism
Voyager doesn't handle failure in ACME challenges well, and moreover starts challenging before the
ingress is ready in the Cloud Provider in the http-01 case. This patch uses the AVI status detection
in the playbook to ensure the Certificate is declared as a resource only when all Ingresses are ready.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>