migration: Ensure we fix broken Astarte resources
When migrating from the old Operator, a number of things have to be fixed due to the fact
that resource validation was beyond flaky. Create a separate migrate package with the sole
purpose of carrying over such migration, and forget about it. Unfortunately, a small downtime
is necessary to go forward with the operation.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>