cfssl-ca-job: Ensure the job is recreated when it fails
Currently, the Operator did not handle the scenario in which the CFSSL CA Job simply fails.
In that case, the job would have remained failed forever, and the CA secret would never be
imported.

With this change, it is detected whether the job failed and, if so, the job is deleted and
will be recreated upon next reconciliation.

Fix #85

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>