AVI: Do not expose /metrics by default
The new monitoring feature in Astarte services exposes a Prometheus-compatible /metrics endpoint.
Such an endpoint, however, is meant to be scraped by a trusted source, and can leak sensitive
information.

Do not serve it by default, and allow fine grained configuration of allowed subnets in case
it should be served to an external scraper.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>