Rationalize project structure
This ensures the controller packages aren't cluttered with the actual reconciliation
logic, which is better kept in a separate structure/package.

So let there be the "lib" package, which contains all the logic.

Also, main.go has been rationalized to ensure future portings to new operator-sdk don't
end up in cluttering the boilerplate.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>