xastarte: Add ManualMaintenanceMode
In some cases, the Operator might fail and manual intervention on the managed resources might
be necessary. In those cases, the Operator is usually shut down, but this doesn't help in
two main cases:

1. The failure happens during an Upgrade (hence the status needs to be computed correctly
   for the upgrade to succeed/resume)
2. The webhooks will not work when the Operator is down, preventing any change to the Astarte
   resource.

ManualMaintenanceMode skips reconciliation entirely while computing status for the resource,
making it possible to do most interventions gracefully. Implemented as a hidden field.

Signed-off-by: Dario Freddi <dario.freddi@ispirata.com>