Support horizontal autoscaling
Extend the Astarte CR to support the k8s HPA using the `autoscaling` feature gate.
The `autoscaler.horizontal` field in an Astarte component's spec allows to
reference an `HorizontalPodAutoscaler` object.
Since queue distribution does not allow for underscaling, RabbitMQ is not autoscalable.
Since DataUpdaterPlant is deployed as a group of independent Deployments, it is not autoscalable.
For persistency reasons, neither the in-cluster Cassandra deployment is autoscalable.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>