fix: use BestEffort QoS for E2E test cr
In an ephemeral env such as E2E testing we allow components to allocate resources as needed without setting requests and limits. This approach is not recommended in prod envs.
For a Pod to be given a QoS class of BestEffort, the Containers in the Pod must not have any memory or CPU limits or requests.  

Signed-off-by: Luca Marchiori <luca.marchiori@secomind.com>