Fix HPA race condition on initialization
When an HPA is created for a resource, it can take some time for the metrics to become available. During this time, the HPA may report a desired replica count of 0, which would cause the Astarte operator to scale down the resource to 0 replicas.

This commit adds a protection mechanism to the operator to prevent this from happening. If the HPA reports 0 desired replicas, the operator will ignore the HPA and use the replica count from the Astarte custom resource instead. This prevents the operator from scaling down the resource to 0 replicas while the HPA is initializing. ISSUE: https://github.com/astarte-platform/astarte-kubernetes-operator/issues/397

Signed-off-by: Guilherme Crocetti <guilherme.crocetti@secomind.com>